add: Player actions using Word2Vec model
This commit implements functionality to get the associated words
for the highlighted word on the screen, using Word2Vec.

The candidate list for the associated words is filtered out based
on multiple factors such as word-length, similarity with original
word, and privious use of the candidate word etc.

Selected candidate word is then entered in the game input field.